convert api.tasks module to package with multiple separate modules for related tasks
