+++ cloudigrade/api/tasks/sources.py
@@ -1,0 +1,337 @@
+"""Tasks for various sources-api operations."""
+import logging
+
+from django.conf import settings
+from django.contrib.auth.models import User
+from django.db import transaction
+from django.db.models import Q
+from django.utils.translation import gettext as _
+from requests.exceptions import BaseHTTPError, RequestException
+
+from api import error_codes
+from api.clouds.aws.tasks import configure_customer_aws_and_create_cloud_account
+from api.clouds.aws.util import update_aws_cloud_account
+from api.models import CloudAccount, UserTaskLock
+from api.tasks.maintenance import _delete_cloud_accounts
+from util import aws
+from util.celery import retriable_shared_task
+from util.exceptions import AwsThrottlingException, KafkaProducerException
+from util.redhatcloud import sources
+
+logger = logging.getLogger(__name__)
+
+
+@retriable_shared_task(
+    autoretry_for=(RequestException, BaseHTTPError, AwsThrottlingException),
+    name="api.tasks.create_from_sources_kafka_message",
+)
+@aws.rewrap_aws_errors
+def create_from_sources_kafka_message(message, headers):
+    """
+    Create our model objects from the Sources Kafka message.
+
+    Because the Sources API may not always be available, this task must
+    gracefully retry if communication with Sources fails unexpectedly.
+
+    If this function succeeds, it spawns another async task to set up the
+    customer's AWS account (configure_customer_aws_and_create_cloud_account).
+
+    Args:
+        message (dict): the "value" attribute of a message from a Kafka
+            topic generated by the Sources service and having event type
+            "ApplicationAuthentication.create"
+        headers (list): the headers of a message from a Kafka topic
+            generated by the Sources service and having event type
+            "ApplicationAuthentication.create"
+
+    """
+    authentication_id = message.get("authentication_id", None)
+    application_id = message.get("application_id", None)
+    (
+        account_number,
+        platform_id,
+    ) = sources.extract_ids_from_kafka_message(message, headers)
+
+    if account_number is None or authentication_id is None or application_id is None:
+        logger.error(_("Aborting creation. Incorrect message details."))
+        return
+
+    application = sources.get_application(account_number, application_id)
+    if not application:
+        logger.info(
+            _(
+                "Application ID %(application_id)s for account number "
+                "%(account_number)s does not exist; aborting cloud account creation."
+            ),
+            {"application_id": application_id, "account_number": account_number},
+        )
+        return
+
+    application_type = application["application_type_id"]
+    if application_type is not sources.get_cloudigrade_application_type_id(
+        account_number
+    ):
+        logger.info(_("Aborting creation. Application Type is not cloudmeter."))
+        return
+
+    authentication = sources.get_authentication(account_number, authentication_id)
+
+    if not authentication:
+        error_code = error_codes.CG2000
+        error_code.log_internal_message(
+            logger,
+            {"authentication_id": authentication_id, "account_number": account_number},
+        )
+        error_code.notify(account_number, application_id)
+        return
+
+    authtype = authentication.get("authtype")
+    if authtype not in settings.SOURCES_CLOUDMETER_AUTHTYPES:
+        error_code = error_codes.CG2001
+        error_code.log_internal_message(
+            logger, {"authentication_id": authentication_id, "authtype": authtype}
+        )
+        error_code.notify(account_number, application_id)
+        return
+
+    resource_type = authentication.get("resource_type")
+    resource_id = authentication.get("resource_id")
+    if resource_type != settings.SOURCES_RESOURCE_TYPE:
+        error_code = error_codes.CG2002
+        error_code.log_internal_message(
+            logger, {"resource_id": resource_id, "account_number": account_number}
+        )
+        error_code.notify(account_number, application_id)
+        return
+
+    source_id = application.get("source_id")
+    arn = authentication.get("username") or authentication.get("password")
+
+    if not arn:
+        error_code = error_codes.CG2004
+        error_code.log_internal_message(
+            logger, {"authentication_id": authentication_id}
+        )
+        error_code.notify(account_number, application_id)
+        return
+
+    with transaction.atomic():
+        user, created = User.objects.get_or_create(username=account_number)
+        if created:
+            user.set_unusable_password()
+            logger.info(
+                _("User %s was not found and has been created."),
+                account_number,
+            )
+            UserTaskLock.objects.get_or_create(user=user)
+
+    # Conditionalize the logic for different cloud providers
+    if authtype == settings.SOURCES_CLOUDMETER_ARN_AUTHTYPE:
+        configure_customer_aws_and_create_cloud_account.delay(
+            user.username,
+            arn,
+            authentication_id,
+            application_id,
+            source_id,
+        )
+
+
+@retriable_shared_task(
+    autoretry_for=(RuntimeError, AwsThrottlingException),
+    name="api.tasks.delete_from_sources_kafka_message",
+)
+@aws.rewrap_aws_errors
+def delete_from_sources_kafka_message(message, headers):
+    """
+    Delete our cloud account as per the Sources Kafka message.
+
+    This function is decorated to retry if an unhandled `RuntimeError` is
+    raised, which is the exception we raise in `rewrap_aws_errors` if we
+    encounter an unexpected error from AWS. This means it should keep retrying
+    if AWS is misbehaving.
+
+    Args:
+        message (dict): a message from the Kafka topic generated by the
+            Sources service and having event type "Authentication.destroy"
+        headers (list): the headers of a message from a Kafka topic
+            generated by the Sources service and having event type
+            "Authentication.destroy" or "Source.destroy"
+
+    """
+    (
+        account_number,
+        platform_id,
+    ) = sources.extract_ids_from_kafka_message(message, headers)
+
+    logger.info(
+        _(
+            "delete_from_sources_kafka_message for account_number %(account_number)s, "
+            "platform_id %(platform_id)s"
+        ),
+        {
+            "account_number": account_number,
+            "platform_id": platform_id,
+        },
+    )
+
+    if account_number is None or platform_id is None:
+        logger.error(_("Aborting deletion. Incorrect message details."))
+        return
+
+    authentication_id = message["authentication_id"]
+    application_id = message["application_id"]
+    query_filter = Q(
+        platform_application_id=application_id,
+        platform_authentication_id=authentication_id,
+    )
+
+    logger.info(_("Deleting CloudAccounts using filter %s"), query_filter)
+    cloud_accounts = CloudAccount.objects.filter(query_filter)
+    _delete_cloud_accounts(cloud_accounts)
+
+
+@retriable_shared_task(
+    autoretry_for=(
+        RequestException,
+        BaseHTTPError,
+        RuntimeError,
+        AwsThrottlingException,
+    ),
+    name="api.tasks.update_from_source_kafka_message",
+)
+@aws.rewrap_aws_errors
+def update_from_source_kafka_message(message, headers):
+    """
+    Update our model objects from the Sources Kafka message.
+
+    Because the Sources API may not always be available, this task must
+    gracefully retry if communication with Sources fails unexpectedly.
+
+    This function is also decorated to retry if an unhandled `RuntimeError` is
+    raised, which is the exception we raise in `rewrap_aws_errors` if we
+    encounter an unexpected error from AWS. This means it should keep retrying
+    if AWS is misbehaving.
+
+    Args:
+        message (dict): the "value" attribute of a message from a Kafka
+            topic generated by the Sources service and having event type
+            "Authentication.update"
+        headers (list): the headers of a message from a Kafka topic
+            generated by the Sources service and having event type
+            "Authentication.update"
+
+    """
+    (
+        account_number,
+        authentication_id,
+    ) = sources.extract_ids_from_kafka_message(message, headers)
+
+    if account_number is None or authentication_id is None:
+        logger.error(_("Aborting update. Incorrect message details."))
+        return
+
+    try:
+        clount = CloudAccount.objects.get(platform_authentication_id=authentication_id)
+
+        authentication = sources.get_authentication(account_number, authentication_id)
+
+        if not authentication:
+            logger.info(
+                _(
+                    "Authentication ID %(authentication_id)s for account number "
+                    "%(account_number)s does not exist; aborting cloud account update."
+                ),
+                {
+                    "authentication_id": authentication_id,
+                    "account_number": account_number,
+                },
+            )
+            return
+
+        resource_type = authentication.get("resource_type")
+        application_id = authentication.get("resource_id")
+        if resource_type != settings.SOURCES_RESOURCE_TYPE:
+            logger.info(
+                _(
+                    "Resource ID %(resource_id)s for account number %(account_number)s "
+                    "is not of type Application; aborting cloud account update."
+                ),
+                {"resource_id": application_id, "account_number": account_number},
+            )
+            return
+
+        application = sources.get_application(account_number, application_id)
+        source_id = application.get("source_id")
+
+        arn = authentication.get("username") or authentication.get("password")
+        if not arn:
+            logger.info(_("Could not update CloudAccount with no ARN provided."))
+            error_code = error_codes.CG2004
+            error_code.log_internal_message(
+                logger, {"authentication_id": authentication_id}
+            )
+            error_code.notify(account_number, application_id)
+            return
+
+        # If the Authentication being updated is arn, do arn things.
+        # The kafka message does not always include authtype, so we get this from
+        # the sources API call
+        if authentication.get("authtype") == settings.SOURCES_CLOUDMETER_ARN_AUTHTYPE:
+            update_aws_cloud_account(
+                clount,
+                arn,
+                account_number,
+                authentication_id,
+                source_id,
+            )
+    except CloudAccount.DoesNotExist:
+        # Is this authentication meant to be for us? We should check.
+        # Get list of all app-auth objects and filter by our authentication
+        response_json = sources.list_application_authentications(
+            account_number, authentication_id
+        )
+
+        if response_json.get("meta").get("count") > 0:
+            for application_authentication in response_json.get("data"):
+                create_from_sources_kafka_message.delay(
+                    application_authentication, headers
+                )
+        else:
+            logger.info(
+                _(
+                    "The updated authentication with ID %s and account number %s "
+                    "is not managed by cloud meter."
+                ),
+                authentication_id,
+                account_number,
+            )
+
+
+@retriable_shared_task(
+    autoretry_for=(KafkaProducerException,),
+    name="api.tasks.notify_application_availability_task",
+)
+def notify_application_availability_task(
+    account_number, application_id, availability_status, availability_status_error=""
+):
+    """
+    Update Sources application's availability status.
+
+    This is a task wrapper to the sources.notify_application_availability
+    method which sends the availability_status Kafka message to Sources.
+
+    Args:
+        account_number (str): Account number identifier
+        application_id (int): Platform insights application id
+        availability_status (string): Availability status to set
+        availability_status_error (string): Optional status error
+    """
+    try:
+        sources.notify_application_availability(
+            account_number,
+            application_id,
+            availability_status,
+            availability_status_error,
+        )
+    except KafkaProducerException:
+        raise
