+++ cloudigrade/api/tests/tasks/calculation/test_fix_problematic_runs.py
@@ -1,10 +1,11 @@
-"""Collection of tests for tasks.fix_problematic_runs."""
+"""Collection of tests for tasks.calculation.fix_problematic_runs."""
 import datetime
 from unittest.mock import call, patch
 
 from django.test import TestCase
 
-from api import models, tasks
+from api import models
+from api.tasks import calculation
 from api.tests import helper as api_helper
 from util.tests import helper as util_helper
 
@@ -12,12 +13,12 @@ from util.tests import helper as util_helper
 class FixProblematicRunsTest(TestCase):
     """Test case for Celery task fix_problematic_runs."""
 
-    @patch("api.tasks._fix_problematic_run")
+    @patch("api.tasks.calculation._fix_problematic_run")
     def test_fix_problematic_runs(self, mock_fix):
         """Test the task simply calls the helper function for each given Run ID."""
         run_ids = [1, 2, 3]
         expected_calls = [call(run_id) for run_id in run_ids]
-        tasks.fix_problematic_runs.run(run_ids)
+        calculation.fix_problematic_runs.run(run_ids)
         mock_fix.assert_has_calls(expected_calls)
 
 
@@ -58,15 +59,15 @@ class FixProblematicRunTest(TestCase):
         self.problematic_run.end_time = None
         self.problematic_run.save()
 
-    @patch("api.tasks.recalculate_runs")
+    @patch("api.tasks.calculation.recalculate_runs")
     def test_run_not_found_does_nothing(self, mock_recalculate):
         """Test early return when the run is not found."""
-        with self.assertLogs("api.tasks", level="INFO") as logging_watcher:
-            tasks._fix_problematic_run(-1)
+        with self.assertLogs("api.tasks.calculation", level="INFO") as logging_watcher:
+            calculation._fix_problematic_run(-1)
         self.assertIn("does not exist", logging_watcher.output[0])
         mock_recalculate.assert_not_called()
 
-    @patch("api.tasks.recalculate_runs")
+    @patch("api.tasks.calculation.recalculate_runs")
     def test_run_found_but_no_power_off_event_does_nothing(self, mock_recalculate):
         """Test early return when the run is found but no power_off event is found."""
         # Start by deleting the power_off event that would be for this run.
@@ -75,24 +76,24 @@ class FixProblematicRunTest(TestCase):
             event_type=models.InstanceEvent.TYPE.power_off
         ).order_by("-occurred_at").first().delete()
 
-        with self.assertLogs("api.tasks", level="INFO") as logging_watcher:
-            tasks._fix_problematic_run(self.problematic_run.id)
+        with self.assertLogs("api.tasks.calculation", level="INFO") as logging_watcher:
+            calculation._fix_problematic_run(self.problematic_run.id)
         self.assertIn("No relevant InstanceEvent exists", logging_watcher.output[0])
         mock_recalculate.assert_not_called()
 
-    @patch("api.tasks.recalculate_runs")
+    @patch("api.tasks.calculation.recalculate_runs")
     def test_run_found_but_is_okay_does_nothing(self, mock_recalculate):
         """Test early return when the run is found but is okay."""
-        with self.assertLogs("api.tasks", level="INFO") as logging_watcher:
-            tasks._fix_problematic_run(self.okay_run.id)
+        with self.assertLogs("api.tasks.calculation", level="INFO") as logging_watcher:
+            calculation._fix_problematic_run(self.okay_run.id)
         self.assertIn("does not appear to require fixing", logging_watcher.output[0])
         mock_recalculate.assert_not_called()
 
-    @patch("api.tasks.recalculate_runs")
+    @patch("api.tasks.calculation.recalculate_runs")
     def test_run_found_and_fixed(self, mock_recalculate):
         """Test attempting to fix when the run is found and should be fixed."""
-        with self.assertLogs("api.tasks", level="INFO") as logging_watcher:
-            tasks._fix_problematic_run(self.problematic_run.id)
+        with self.assertLogs("api.tasks.calculation", level="INFO") as logging_watcher:
+            calculation._fix_problematic_run(self.problematic_run.id)
         self.assertIn("Attempting to fix problematic runs", logging_watcher.output[0])
         last_power_off = (
             models.InstanceEvent.objects.filter(
