+++ cloudigrade/api/tests/tasks/sources/test_create_from_sources_kafka_message.py
@@ -6,7 +6,7 @@ from django.conf import settings
 from django.contrib.auth.models import User
 from django.test import TestCase
 
-from api import tasks
+from api.tasks import sources
 from util.exceptions import SourcesAPINotOkStatus
 from util.tests import helper as util_helper
 
@@ -46,7 +46,7 @@ class CreateFromSourcesKafkaMessageTest(TestCase):
     @patch("util.redhatcloud.sources.get_authentication")
     @patch("util.redhatcloud.sources.get_cloudigrade_application_type_id")
     @patch("util.redhatcloud.sources.get_application")
-    @patch("api.tasks.configure_customer_aws_and_create_cloud_account")
+    @patch("api.tasks.sources.configure_customer_aws_and_create_cloud_account")
     def test_create_from_sources_kafka_message_success(
         self,
         mock_task,
@@ -65,7 +65,7 @@ class CreateFromSourcesKafkaMessageTest(TestCase):
 
         self.auth_return_value["username"] = arn
         mock_get_auth.return_value = self.auth_return_value
-        tasks.create_from_sources_kafka_message(self.message, self.headers)
+        sources.create_from_sources_kafka_message(self.message, self.headers)
 
         user = User.objects.get(username=self.account_number)
         mock_task.delay.assert_called_with(
@@ -77,34 +77,34 @@ class CreateFromSourcesKafkaMessageTest(TestCase):
         )
 
     @patch("util.redhatcloud.sources.get_application")
-    @patch("api.tasks.configure_customer_aws_and_create_cloud_account")
+    @patch("api.tasks.sources.configure_customer_aws_and_create_cloud_account")
     def test_aws_task_not_called_for_missing_application_id(
         self, mock_task, mock_get_application
     ):
         """Assert early exit if application_id not in message."""
         self.message["application_id"] = None
-        tasks.create_from_sources_kafka_message(self.message, self.headers)
+        sources.create_from_sources_kafka_message(self.message, self.headers)
         self.assertEqual(User.objects.all().count(), 0)
         mock_task.delay.assert_not_called()
         mock_get_application.assert_not_called()
 
     @patch("util.redhatcloud.sources.get_application")
-    @patch("api.tasks.configure_customer_aws_and_create_cloud_account")
+    @patch("api.tasks.sources.configure_customer_aws_and_create_cloud_account")
     def test_aws_task_not_called_for_missing_authentication_id(
         self, mock_task, mock_get_application
     ):
         """Assert early exit if authentication_id not in message."""
         self.message["authentication_id"] = None
-        tasks.create_from_sources_kafka_message(self.message, self.headers)
+        sources.create_from_sources_kafka_message(self.message, self.headers)
         self.assertEqual(User.objects.all().count(), 0)
         mock_task.delay.assert_not_called()
         mock_get_application.assert_not_called()
 
-    @patch("api.tasks.notify_application_availability_task")
+    @patch("api.tasks.sources.notify_application_availability_task")
     @patch("util.redhatcloud.sources.get_authentication")
     @patch("util.redhatcloud.sources.get_cloudigrade_application_type_id")
     @patch("util.redhatcloud.sources.get_application")
-    @patch("api.tasks.configure_customer_aws_and_create_cloud_account")
+    @patch("api.tasks.sources.configure_customer_aws_and_create_cloud_account")
     def test_aws_task_not_called_for_unsupported_authtype(
         self,
         mock_task,
@@ -120,11 +120,11 @@ class CreateFromSourcesKafkaMessageTest(TestCase):
         mock_get_app_type_id.return_value = self.cloudigrade_sources_app_id
         self.auth_return_value["authtype"] = "INVALID"
         mock_get_auth.return_value = self.auth_return_value
-        tasks.create_from_sources_kafka_message(self.message, self.headers)
+        sources.create_from_sources_kafka_message(self.message, self.headers)
         mock_task.delay.assert_not_called()
 
     @patch("util.redhatcloud.sources.get_application")
-    @patch("api.tasks.configure_customer_aws_and_create_cloud_account")
+    @patch("api.tasks.sources.configure_customer_aws_and_create_cloud_account")
     def test_create_from_sources_kafka_message_fail_source_unexpected_response(
         self, mock_task, mock_get_app
     ):
@@ -136,14 +136,14 @@ class CreateFromSourcesKafkaMessageTest(TestCase):
         mock_get_app.side_effect = SourcesAPINotOkStatus
 
         with self.assertRaises(SourcesAPINotOkStatus):
-            tasks.create_from_sources_kafka_message(self.message, self.headers)
+            sources.create_from_sources_kafka_message(self.message, self.headers)
 
         # User should not have been created.
         self.assertEqual(User.objects.all().count(), 0)
         mock_task.delay.assert_not_called()
 
     @patch("util.redhatcloud.sources.get_application")
-    @patch("api.tasks.configure_customer_aws_and_create_cloud_account")
+    @patch("api.tasks.sources.configure_customer_aws_and_create_cloud_account")
     def test_create_from_sources_kafka_message_returns_early_when_application_404(
         self, mock_task, mock_get_app
     ):
@@ -155,17 +155,17 @@ class CreateFromSourcesKafkaMessageTest(TestCase):
         """
         mock_get_app.return_value = None
 
-        tasks.create_from_sources_kafka_message(self.message, self.headers)
+        sources.create_from_sources_kafka_message(self.message, self.headers)
 
         # User should not have been created.
         self.assertEqual(User.objects.all().count(), 0)
         mock_task.delay.assert_not_called()
 
-    @patch("api.tasks.notify_application_availability_task")
+    @patch("api.tasks.sources.notify_application_availability_task")
     @patch("util.redhatcloud.sources.get_authentication")
     @patch("util.redhatcloud.sources.get_cloudigrade_application_type_id")
     @patch("util.redhatcloud.sources.get_application")
-    @patch("api.tasks.configure_customer_aws_and_create_cloud_account")
+    @patch("api.tasks.sources.configure_customer_aws_and_create_cloud_account")
     def test_create_from_sources_kafka_message_returns_early_when_authentication_404(
         self,
         mock_task,
@@ -185,17 +185,17 @@ class CreateFromSourcesKafkaMessageTest(TestCase):
         }
         mock_get_app_type_id.return_value = self.cloudigrade_sources_app_id
         mock_get_auth.return_value = None
-        tasks.create_from_sources_kafka_message(self.message, self.headers)
+        sources.create_from_sources_kafka_message(self.message, self.headers)
 
         # User should not have been created.
         self.assertEqual(User.objects.all().count(), 0)
         mock_task.delay.assert_not_called()
 
-    @patch("api.tasks.notify_application_availability_task")
+    @patch("api.tasks.sources.notify_application_availability_task")
     @patch("util.redhatcloud.sources.get_authentication")
     @patch("util.redhatcloud.sources.get_cloudigrade_application_type_id")
     @patch("util.redhatcloud.sources.get_application")
-    @patch("api.tasks.configure_customer_aws_and_create_cloud_account")
+    @patch("api.tasks.sources.configure_customer_aws_and_create_cloud_account")
     def test_create_returns_early_when_resource_type_invalid(
         self,
         mock_task,
@@ -213,16 +213,16 @@ class CreateFromSourcesKafkaMessageTest(TestCase):
         self.auth_return_value["resource_type"] = "INVALID"
         mock_get_auth.return_value = self.auth_return_value
 
-        tasks.create_from_sources_kafka_message(self.message, self.headers)
+        sources.create_from_sources_kafka_message(self.message, self.headers)
         # User should not have been created.
         self.assertEqual(User.objects.all().count(), 0)
         mock_task.delay.assert_not_called()
 
-    @patch("api.tasks.notify_application_availability_task")
+    @patch("api.tasks.sources.notify_application_availability_task")
     @patch("util.redhatcloud.sources.get_authentication")
     @patch("util.redhatcloud.sources.get_cloudigrade_application_type_id")
     @patch("util.redhatcloud.sources.get_application")
-    @patch("api.tasks.configure_customer_aws_and_create_cloud_account")
+    @patch("api.tasks.sources.configure_customer_aws_and_create_cloud_account")
     def test_create_fails_if_no_password_or_username(
         self,
         mock_task,
@@ -240,14 +240,14 @@ class CreateFromSourcesKafkaMessageTest(TestCase):
         self.auth_return_value.pop("username", None)
         self.auth_return_value.pop("password", None)
         mock_get_auth.return_value = self.auth_return_value
-        tasks.create_from_sources_kafka_message(self.message, self.headers)
+        sources.create_from_sources_kafka_message(self.message, self.headers)
 
         mock_task.delay.assert_not_called()
 
     @patch("util.redhatcloud.sources.get_authentication")
     @patch("util.redhatcloud.sources.get_cloudigrade_application_type_id")
     @patch("util.redhatcloud.sources.get_application")
-    @patch("api.tasks.configure_customer_aws_and_create_cloud_account")
+    @patch("api.tasks.sources.configure_customer_aws_and_create_cloud_account")
     def test_early_exit_if_app_type_is_not_cloudigrade(
         self, mock_task, mock_get_app, mock_get_app_type_id, mock_get_auth
     ):
@@ -255,7 +255,7 @@ class CreateFromSourcesKafkaMessageTest(TestCase):
         mock_get_app.return_value = {"application_type_id": "INVALID"}
         mock_get_app_type_id.return_value = self.cloudigrade_sources_app_id
 
-        tasks.create_from_sources_kafka_message(self.message, self.headers)
+        sources.create_from_sources_kafka_message(self.message, self.headers)
         mock_get_auth.assert_not_called()
 
         mock_task.delay.assert_not_called()
@@ -263,7 +263,7 @@ class CreateFromSourcesKafkaMessageTest(TestCase):
     @patch("util.redhatcloud.sources.get_authentication")
     @patch("util.redhatcloud.sources.get_cloudigrade_application_type_id")
     @patch("util.redhatcloud.sources.get_application")
-    @patch("api.tasks.configure_customer_aws_and_create_cloud_account")
+    @patch("api.tasks.sources.configure_customer_aws_and_create_cloud_account")
     def test_create_uses_arn_from_password_if_no_username(
         self,
         mock_task,
@@ -283,7 +283,7 @@ class CreateFromSourcesKafkaMessageTest(TestCase):
         self.auth_return_value["password"] = arn
 
         mock_get_auth.return_value = self.auth_return_value
-        tasks.create_from_sources_kafka_message(self.message, self.headers)
+        sources.create_from_sources_kafka_message(self.message, self.headers)
 
         user = User.objects.get(username=self.account_number)
         mock_task.delay.assert_called_with(
