+++ cloudigrade/api/tests/test_models.py
@@ -112,7 +112,7 @@ class RunTest(TestCase):
 class CloudAccountTest(TestCase):
     """Test cases for api.models.CloudAccount."""
 
-    @patch("api.tasks.notify_application_availability_task")
+    @patch("api.tasks.sources.notify_application_availability_task")
     def test_delete_clount_deletes_user(self, mock_notify_sources):
         """Test User is deleted if last clount is deleted."""
         user = util_helper.generate_test_user()
@@ -126,7 +126,7 @@ class CloudAccountTest(TestCase):
         account.delete()
         self.assertFalse(User.objects.filter(username=username).exists())
 
-    @patch("api.tasks.notify_application_availability_task")
+    @patch("api.tasks.sources.notify_application_availability_task")
     def test_delete_clount_doesnt_delete_user_for_two_clounts(
         self, mock_notify_sources
     ):
