+++ cloudigrade/internal/tests/views/test_recalculate_runs.py
@@ -23,7 +23,7 @@ class RecalculateRunsViewTest(TestCase):
         )
 
     @override_settings(CELERY_TASK_ALWAYS_EAGER=True)
-    @patch("api.tasks._recalculate_runs_for_cloud_account_id")
+    @patch("api.tasks.calculation._recalculate_runs_for_cloud_account_id")
     def test_recalculate_runs_no_args(self, mock_recalculate):
         """Assert recalculate task triggered with no args."""
         request = self.factory.post("/recalculate_runs/", data={}, format="json")
@@ -32,7 +32,7 @@ class RecalculateRunsViewTest(TestCase):
         mock_recalculate.assert_called_with(self.account.id, None)
 
     @override_settings(CELERY_TASK_ALWAYS_EAGER=True)
-    @patch("api.tasks._recalculate_runs_for_cloud_account_id")
+    @patch("api.tasks.calculation._recalculate_runs_for_cloud_account_id")
     def test_recalculate_runs_cloud_account_id(self, mock_recalculate):
         """Assert recalculate task triggered with valid cloud_account_id."""
         request = self.factory.post(
@@ -43,7 +43,7 @@ class RecalculateRunsViewTest(TestCase):
         mock_recalculate.assert_called_with(420, None)
 
     @override_settings(CELERY_TASK_ALWAYS_EAGER=True)
-    @patch("api.tasks._recalculate_runs_for_cloud_account_id")
+    @patch("api.tasks.calculation._recalculate_runs_for_cloud_account_id")
     def test_recalculate_runs_bad_cloud_account_id(self, mock_recalculate):
         """Assert recalculate task not triggered with malformed cloud_account_id."""
         request = self.factory.post(
@@ -54,7 +54,7 @@ class RecalculateRunsViewTest(TestCase):
         mock_recalculate.assert_not_called()
 
     @override_settings(CELERY_TASK_ALWAYS_EAGER=True)
-    @patch("api.tasks._recalculate_runs_for_cloud_account_id")
+    @patch("api.tasks.calculation._recalculate_runs_for_cloud_account_id")
     def test_recalculate_runs_since(self, mock_recalculate):
         """Assert recalculate task triggered with valid since date."""
         request = self.factory.post(
@@ -66,7 +66,7 @@ class RecalculateRunsViewTest(TestCase):
         mock_recalculate.assert_called_with(self.account.id, expected_since)
 
     @override_settings(CELERY_TASK_ALWAYS_EAGER=True)
-    @patch("api.tasks._recalculate_runs_for_cloud_account_id")
+    @patch("api.tasks.calculation._recalculate_runs_for_cloud_account_id")
     def test_recalculate_runs_since_with_timezone(self, mock_recalculate):
         """Assert recalculate task triggered with valid since datetime+offset."""
         request = self.factory.post(
@@ -82,7 +82,7 @@ class RecalculateRunsViewTest(TestCase):
         mock_recalculate.assert_called_with(self.account.id, expected_since)
 
     @override_settings(CELERY_TASK_ALWAYS_EAGER=True)
-    @patch("api.tasks._recalculate_runs_for_cloud_account_id")
+    @patch("api.tasks.calculation._recalculate_runs_for_cloud_account_id")
     def test_recalculate_runs_bad_since(self, mock_recalculate):
         """Assert recalculate task not triggered with malformed since."""
         request = self.factory.post(
