+++ cloudigrade/util/tests/test_misc.py
@@ -68,7 +68,7 @@ class LockUserTaskTest(TransactionTestCase):
         self.assertEqual(lock.locked, False)
 
     @patch("api.clouds.aws.util.delete_cloudtrail")
-    @patch("api.tasks.notify_application_availability_task")
+    @patch("api.tasks.sources.notify_application_availability_task")
     def test_delete_clount_lock(self, mock_notify_sources, mock_delete_cloudtrail):
         """Test that deleting an clount inside a lock is successful."""
         aws_account_id = util_helper.generate_dummy_aws_account_id()
@@ -86,7 +86,7 @@ class LockUserTaskTest(TransactionTestCase):
         self.assertEqual(CloudAccount.objects.all().count(), 0)
         self.assertFalse(UserTaskLock.objects.filter(user=account.user).exists())
 
-    @patch("api.tasks.notify_application_availability_task")
+    @patch("api.tasks.sources.notify_application_availability_task")
     def test_delete_clount_lock_exception(self, mock_notify_sources):
         """Test that an exception when deleting a clount inside a lock rolls back."""
         aws_account_id = util_helper.generate_dummy_aws_account_id()
